requirements.txt
fastapi
pillow
pytest
httpx
python-multipart
